Change runtime call param count

+++ src/hotspot/share/runtime/sharedRuntime.hpp
@@ -161,14 +161,14 @@ class SharedRuntime: AllStatic {
                                 jint len, jlong inv, jint *m_ints);
 
   // oddModPow
-  static void oddModPowInner1K(jint *base, jint *exp, jint *modulus, jlong modLen,
-                               jint *toMont, jlong montLen, jlong inv, jint *result);
+  static void oddModPowInner1K(jint *base, jint *exp, jint *modulus, jint modLen,
+                               jint *toMont, jint montLen, jlong inv, jint *result);
 
-  static void oddModPowInner1o5K(jint *base, jint *exp, jint *modulus, jlong modLen,
-                               jint *toMont, jlong montLen, jlong inv, jint *result);
+  static void oddModPowInner1o5K(jint *base, jint *exp, jint *modulus, jint modLen,
+                               jint *toMont, jint montLen, jlong inv, jint *result);
 
-  static void oddModPowInner2K(jint *base, jint *exp, jint *modulus, jlong modLen,
-                               jint *toMont, jlong montLen, jlong inv, jint *result);
+  static void oddModPowInner2K(jint *base, jint *exp, jint *modulus, jint modLen,
+                               jint *toMont, jint montLen, jlong inv, jint *result);
 
 #ifdef __SOFTFP__
   // C++ compiler generates soft float instructions as well as passing
